+++ src/main/kotlin/com/insyncwithfoo/pyrightls/PyrightLSLspServerDescriptor.kt
@@ -1,4 +1,4 @@
-package com.insyncwithfoo.pyrightexperimental
+package com.insyncwithfoo.pyrightls
 
 import com.intellij.execution.configurations.GeneralCommandLine
 import com.intellij.openapi.project.Project
@@ -8,19 +8,22 @@ import java.io.File
 
 
 @Suppress("UnstableApiUsage")
-class PyrightLspServerDescriptor(
+class PyrightLSLspServerDescriptor(
     project: Project,
     private val executable: File
 ) : ProjectWideLspServerDescriptor(project, PRESENTABLE_NAME) {
-
+    
     override fun isSupportedFile(file: VirtualFile) = file.extension == "py"
-
-    override fun createCommandLine() = GeneralCommandLine(executable.canonicalPath, "--stdio")
-
-    override val lspServerListener = PyrightLspServerListener(project)
-
+    
+    override fun createCommandLine() =
+        GeneralCommandLine(executable.canonicalPath, "--stdio").apply {
+            withCharset(Charsets.UTF_8)
+        }
+    
+    override val lspServerListener = PyrightLSLspServerListener(project)
+    
     companion object {
         const val PRESENTABLE_NAME = "Pyright LS"
     }
-
+    
 }