+++ src/main/kotlin/com/insyncwithfoo/pyrightls/configuration/application/ConfigurationService.kt
@@ -1,17 +1,20 @@
-package com.insyncwithfoo.pyrightexperimental.configuration.project
+package com.insyncwithfoo.pyrightls.configuration.application
 
-import com.insyncwithfoo.pyrightexperimental.configuration.HasConfigurations
+import com.insyncwithfoo.pyrightls.configuration.HasConfigurations
+import com.intellij.openapi.components.RoamingType
 import com.intellij.openapi.components.Service
 import com.intellij.openapi.components.SimplePersistentStateComponent
 import com.intellij.openapi.components.State
 import com.intellij.openapi.components.Storage
 import com.intellij.openapi.components.service
-import com.intellij.openapi.project.Project
 import com.intellij.util.xmlb.XmlSerializerUtil
 
 
-@State(name = "ProjectConfigurations", storages = [Storage("pyright-experimental.xml")])
-@Service(Service.Level.PROJECT)
+@State(
+    name = "ApplicationConfigurations",
+    storages = [Storage("pyright-langserver.xml", roamingType = RoamingType.DISABLED)]
+)
+@Service(Service.Level.APP)
 class ConfigurationService :
     SimplePersistentStateComponent<Configurations>(Configurations()),
     HasConfigurations<Configurations> {
@@ -21,7 +24,7 @@ class ConfigurationService :
         set(value) = XmlSerializerUtil.copyBean(value, state)
     
     companion object {
-        fun getInstance(project: Project): ConfigurationService = project.service()
+        fun getInstance() = service<ConfigurationService>()
     }
     
 }